+++ src/refactoring/components/admin/coupon/index.ts
@@ -1,0 +1,2 @@
+export * from "./CouponForm";
+export * from "./CouponList";
