+++ src/refactoring/components/admin/product/index.ts
@@ -1,0 +1,5 @@
+export * from "./AdminProductInfo";
+export * from "./DiscountInfo";
+export * from "./ProductList";
+export * from "./ProductToggleButton";
+export * from "./UpdateProductForm";
