+++ src/refactoring/pages/admin.tsx
@@ -1,5 +1,5 @@
 import type { Product, CouponItem } from "../types";
-import { AdminProductList, AdminCouponList } from "../components";
+import { AdminProductInfo, AdminCouponInfo } from "../components";
 interface AdminPageProps {
   products: Product[];
   coupons: CouponItem[];
@@ -18,12 +18,12 @@ const AdminPage = ({
     <div className="container mx-auto p-4">
       <h1 className="text-3xl font-bold mb-6">관리자 페이지</h1>
       <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-        <AdminProductList
+        <AdminProductInfo
           products={products}
           onProductUpdate={onProductUpdate}
           onProductAdd={onProductAdd}
         />
-        <AdminCouponList coupons={coupons} onCouponAdd={onCouponAdd} />
+        <AdminCouponInfo coupons={coupons} onCouponAdd={onCouponAdd} />
       </div>
     </div>
   );
